refactor: rename useSignIn.tsx to use-sign-in.tsx (#29)
