+++ frontend/src/features/auth/components/SignInForm.tsx
@@ -1,3 +1,4 @@
+import { useSignIn } from '../hooks/use-sign-in';
 import EmailInput from './EmailInput';
 import ErrorMessage from './ErrorMessage';
 import PasswordInput from './PasswordInput';
