+++ frontend/src/features/auth/components/SignUpForm.tsx
@@ -1,3 +1,4 @@
+import { useSignIn } from '../hooks/use-sign-in';
 import EmailInput from './EmailInput';
 import ErrorMessage from './ErrorMessage';
 import PasswordInput from './PasswordInput';
