+++ frontend/src/features/auth/components/SubmitButton.tsx
@@ -1,6 +1,7 @@
 type SubmitButtonProps = {
   isSubmitting: boolean;
 };
+import { useSignIn } from '../hooks/use-sign-in';
 
 export default function SubmitButton({ isSubmitting }: SubmitButtonProps) {
   return (
