+++ frontend/src/features/auth/index.ts
@@ -1,6 +1,6 @@
 export { AuthProvider } from './contexts/AuthContext';
-export { useSignIn } from './hooks/useSignIn';
-export { default as SignInPage } from './pages/SignInPage';
+export { useSignIn } from './hooks/use-sign-in';
+export { default as AuthPage } from './pages/AuthPage';
 export { default as GuestRoute } from './routes/GuestRoute';
 export { default as ProtectedRoute } from './routes/ProtectedRoute';
 export * from './types/auth';
